Add functions to types
