+++ account.go
@@ -1,9 +1,10 @@
 package substate
 
 import (
+	"bytes"
 	"math/big"
 
-	"github.com/ethereum/go-ethereum/common"
+	"github.com/Fantom-foundation/Substate/common"
 )
 
 // Account holds any information about account used in a transaction.
@@ -14,11 +15,56 @@ type Account struct {
 	Code    []byte
 }
 
-func NewAccount(nonce uint64, balance *big.Int, storage map[common.Hash]common.Hash, code []byte) *Account {
+func NewAccount(nonce uint64, balance *big.Int, code []byte) *Account {
 	return &Account{
 		Nonce:   nonce,
 		Balance: balance,
-		Storage: storage,
 		Code:    code,
 	}
 }
+
+// Equal returns true if a is y or if values of a are equal to values of y.
+// Otherwise, a and y are not equal hence false is returned.
+func (a *Account) Equal(y *Account) bool {
+	if a == y {
+		return true
+	}
+
+	if (a == nil || y == nil) && a != y {
+		return false
+	}
+
+	// check values
+	equal := a.Nonce == y.Nonce &&
+		a.Balance.Cmp(y.Balance) == 0 &&
+		bytes.Equal(a.Code, y.Code) &&
+		len(a.Storage) == len(y.Storage)
+	if !equal {
+		return false
+	}
+
+	for aKey, aVal := range a.Storage {
+		yValue, exist := y.Storage[aKey]
+		if !(exist && aVal == yValue) {
+			return false
+		}
+	}
+
+	return true
+}
+
+// Copy returns a hard copy of a
+func (a *Account) Copy() *Account {
+	cpy := NewAccount(a.Nonce, a.Balance, a.Code)
+
+	for key, value := range a.Storage {
+		cpy.Storage[key] = value
+	}
+
+	return cpy
+}
+
+// CodeHash returns hashed code
+//func (a *Account) CodeHash() common.Hash { todo uncomment when eth is copied
+//	return crypto.Keccak256Hash(a.Code)
+//}
