+++ result.go
@@ -1,6 +1,8 @@
 package substate
 
 import (
+	"bytes"
+
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 )
@@ -23,4 +25,45 @@ func NewResult(status uint64, bloom types.Bloom, logs []*types.Log, contractAddr
 		ContractAddress: contractAddress,
 		GasUsed:         gasUsed,
 	}
+}
+
+// Equal returns true if r is y or if values of r are equal to values of y.
+// Otherwise, r and y are not equal hence false is returned.
+func (r *Result) Equal(y *Result) bool {
+	if r == y {
+		return true
+	}
+
+	if (r == nil || y == nil) && r != y {
+		return false
+	}
+
+	equal := r.Status == y.Status &&
+		r.Bloom == y.Bloom &&
+		len(r.Logs) == len(y.Logs) &&
+		r.ContractAddress == y.ContractAddress &&
+		r.GasUsed == y.GasUsed
+	if !equal {
+		return false
+	}
+
+	for i, logs := range r.Logs {
+		yLogs := y.Logs[i]
+
+		equal := logs.Address == yLogs.Address &&
+			len(logs.Topics) == len(yLogs.Topics) &&
+			bytes.Equal(logs.Data, yLogs.Data)
+		if !equal {
+			return false
+		}
+
+		for i, xt := range logs.Topics {
+			yt := yLogs.Topics[i]
+			if xt != yt {
+				return false
+			}
+		}
+	}
+
+	return true
 }
